Popup

+++ src/components/table/Table.js
@@ -1,23 +1,22 @@
 import {useState, useContext} from "react";
-
-import { WordsContext } from "../../context/wordsContext";
-import { SelectContext } from "../../context/selectContext";
-
 import classnames from 'classnames';
+
+import { WordsContext } from "../../context/WordsContext";
 
 import Row from '../row/Row';
 import Select from "../select/Select";
 import Pagination from "../pagination/Pagination";
-import Modal from "../modal/Modal";
+import useConfirm from "../../hooks/useConfirm";
 
 import './table.scss';
 
 
 const Table = () => {
 
+    const {confirm} = useConfirm();
+
     const {words, deleteWords} = useContext(WordsContext);
     const [wordList] = useState(words);
-    const {term} = useContext(SelectContext);
 
     const {currentPage, setCurrentPage} = useContext(WordsContext);
     const [wordsPerPage] = useState(7);
@@ -25,26 +24,31 @@ const Table = () => {
     const indexOfLastWord = currentPage * wordsPerPage;
     const indexOfFirstWord = indexOfLastWord - wordsPerPage;
 
-    const onDelete = (id) => {
-        deleteWords(id);
+    const {term, setTerm} = useContext(WordsContext);
+    const [selectVisible, setSelectVisible] = useState(false);
+
+    const showConfirm = async (id) => {
+        const isConfirmed = await confirm('Удалить слово?');
+
+        if (isConfirmed) {
+            deleteWords(id);
+            setTerm('');
+        } else {
+            setTerm('');
+        }
     }
 
     const elements = wordList.map(word => (
         <Row
             key={word.id}
             {...word}
-            onDelete={onDelete}/>
+            onDelete={showConfirm}
+            />
         ))
 
-    //****НОВЫЙ КОД:НАЧАЛО****
-    //Фильтруем
     const filteredElements = term ==='' ? elements : elements.filter(element => element.props.tags === term);
-    //Разбиваем на страницы
     const currentWords = filteredElements.slice(indexOfFirstWord, indexOfLastWord);
-    //****НОВЫЙ КОД:КОНЕЦ****
-
     const paginate = (pageNumber) => setCurrentPage(pageNumber);
-    const [selectVisible, setSelectVisible] = useState(false);
 
     const openSelectIconClasses = classnames({
         'showSelectIcon': true,
@@ -56,9 +60,7 @@ const Table = () => {
         setSelectVisible(!selectVisible);
     }
 
-
     return (
-        <>
         <table className="app__table table">
             <thead>
                 <tr className="table__row">
@@ -66,7 +68,10 @@ const Table = () => {
                     <th>English</th>
                     <th>Transcription</th>
                     <th>Russian</th>
-                    <th className='collection'>Collection <i className={openSelectIconClasses} onClick={onCaretClick}></i>{selectVisible && <Select/>}</th>
+                    <th className='collection'>Collection
+                        <i className={openSelectIconClasses} onClick={onCaretClick}/>
+                        {selectVisible && <Select/>}
+                    </th>
                     <th>Action</th>
                 </tr>
             </thead>
@@ -85,8 +90,6 @@ const Table = () => {
                 </tr>
             </tfoot>
         </table>
-        <Modal/>
-        </>
     )
 }
